feat(docs): generate docs for non react modules
