+++ packages/docs/docgen.js
@@ -4,44 +4,54 @@ const reactDocs = require('react-docgen');
 const glob = require('glob');
 const fse = require('fs-extra');
 const chokidar = require('chokidar');
+const createJsdocContent = require('./functions-generator');
 
 const COMPONENTS_JSON = 'component-docs.json';
 
 const componentsSrc = path.resolve(__dirname, '../*/src');
-const files = glob.sync(`${componentsSrc}/**/*.js`);
+const files = glob.sync(`${componentsSrc}/**/*.js`).filter(file => !file.match(/((test|spec|index).js|(\/__mock__\/|\/__mocks__\/|\/test\/))/gmi));
 
 const args = process.argv.slice(2);
 
 let content = {};
-function parseFile(file, content) {
+async function parseFile(file, content) {
     const src = fse.readFileSync(file, { encoding: 'utf-8' });
     try {
         const componentInfo = reactDocs.parse(src);
         content[file] = componentInfo;
     } catch (error) {
-        console.log('\x1b[33m%s\x1b[0m', error.message, `File: ${file}`);
+        if (error.message.includes('No suitable component definition found')) {
+            const jsdocContent = await createJsdocContent(file);
+            if (jsdocContent) {
+                content[file] = jsdocContent;
+            }
+
+        } else {
+            console.log('\x1b[33m%s\x1b[0m', error.message, `File: ${file}`);
+        }
     }
 }
 
-const target = path.resolve(__dirname, `./${COMPONENTS_JSON}`);
+const componentsTarget = path.resolve(__dirname, `./${COMPONENTS_JSON}`);
 
-function run(files) {
+async function run(files) {
     content = {};
-    files.forEach(file => {
-        parseFile(file, content);
+    const cmds = files.map(file => {
+        return parseFile(file, content);
     });
-    fse.writeJSONSync(target, content, { spaces: '\t' });
+    await Promise.all(cmds);
+    fse.writeJSONSync(componentsTarget, content, { spaces: '\t' });
 }
 
 console.log('Generating react-docgen JSON\n');
 if (args.includes('-w') || args.includes('--watch')) {
     console.log(`Watching ${files.length} JS files in: ${path.resolve(__dirname)}`);
     const watcher = chokidar.watch(files);
-    watcher.on('change', path => {
+    watcher.on('change', async path => {
         console.log(`Compiling file: ${path}`);
-        const currentContent = fse.readJSONSync(target);
-        parseFile(path, currentContent);
-        fse.writeJSONSync(target, currentContent, { spaces: '\t' });
+        const currentContent = fse.readJSONSync(componentsTarget);
+        await parseFile(path, currentContent);
+        fse.writeJSONSync(componentsTarget, currentContent, { spaces: '\t' });
     });
 
 } else {
