+++ packages/docs/generate-components-nav.js
@@ -9,7 +9,7 @@ function getComponentNav(path) {
     };
 }
 
-function assignComponentToNav(nav, { packageName, groupName, navName }) {
+function assignComponentToNav(nav, { packageName, groupName, navName, title }) {
     if (!nav[packageName]) {
         nav[packageName] = {};
     }
@@ -18,15 +18,21 @@ function assignComponentToNav(nav, { packageName, groupName, navName }) {
         nav[packageName][groupName] = [];
     }
 
-    nav[packageName][groupName].push(navName);
+    nav[packageName][groupName].push(title ? { title, name: navName } : navName);
 }
 
 function generateComponentsNavigation(components) {
     const nav = {};
-    Object.keys(components).forEach(path => {
+    Object.entries(components).forEach(([ path, API ]) => {
         const packageName = sortFile(path);
         const { navName, groupName } = getComponentNav(path);
-        assignComponentToNav(nav, { packageName, navName, groupName });
+        if (API.jsdoc) {
+            API.items.forEach(({ name }) => {
+                assignComponentToNav(nav, { packageName, navName: `${navName}-${name}`, groupName, title: name });
+            });
+        } else {
+            assignComponentToNav(nav, { packageName, navName, groupName });
+        }
     });
     return Object.entries(nav).map(([ key, groups ]) => ({
         packageName: key,
