+++ packages/docs/package.json
@@ -31,6 +31,8 @@
     "codesandbox": "^2.2.1",
     "concurrently": "^5.3.0",
     "file-loader": "^6.2.0",
+    "jsdoc": "^3.6.6",
+    "jsdoc3-parser": "^2.0.0",
     "next-transpile-modules": "^6.0.0",
     "raw-loader": "^4.0.2",
     "react-docgen": "^5.3.1",
